Splits forummodel.class.php
