+++ lib/controller/api/ForumController.class.php
@@ -1,6 +1,6 @@
 <?php
 
-use \Jacwright\RestServer\RestException;
+use Jacwright\RestServer\RestException;
 
 class ApiForumController {
